Archive weaviatest.
The weaviatest project has been moved into weaviate-cli, as the
tool grew from more than a testing framework into a tool that
allows managing weaviate instances we felt there was a need for
renaming it, as well as improving its mantainability. That is the
reason why we rescued weaviate-cli and turned it into a better
weaviatest version.